aider: Define `HEAD` as 0, the index of the snake's head in the `wormCoords` list.
# Context:
USER: NameError: name 'HEAD' is not defined
ASSISTANT: The error message is indicating that the `HEAD` variable is not defined in your code. In the context of your code, `HEAD` is used to refer to the first segment of the snake. Let's define `HEAD` as 0 at the top of your `main.py` file.

```python
new_file/main.py
<<<<<<< ORIGINAL
# set up the direction variables
UP = 'up'
DOWN = 'down'
LEFT = 'left'
RIGHT = 'right'
=======
# set up the direction variables
UP = 'up'
DOWN = 'down'
LEFT = 'left'
RIGHT = 'right'
HEAD = 0 # index of the snake's head
>>>>>>> UPDATED
```

This will define `HEAD` as 0, which is the index of the snake's head in the `wormCoords` list.